Changed to Qt 6.6 Improved picker for highdpi screens

+++ colorwheel.cpp
@@ -244,7 +244,7 @@ ColorwheelPrivate::update()
                 {
                     QTransform world = p.worldTransform();
                     states.push_back(State {
-                        states.count(),
+                        (int)states.count(),
                         color,
                         world.mapRect(rect)
                     });
